Update datastore.js
Contrary to what is stated in the doc, updatedAt is always set to new Date() on document update, even if an updatedAt value is present in the request. I added an 'if' statement that changes that behaviour and keeps the request updatedAt value if present (this is necessary for synchronization mechanism).